requirements.txt
appdirs==1.4.3
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.4.0.2
bleach==1.5.0
configparser==3.5.0
cycler==0.10.0
decorator==4.0.11
docopt==0.6.2
entrypoints==0.2.2
enum34==1.1.6
funcsigs==1.0.2
functools32==3.2.3.post2
gitdb2==2.0.0
GitPython==2.1.3
h5py==2.6.0
html5lib==0.9999999
ipdb==0.10.2
ipykernel==4.5.2
ipython==5.2.2
ipython-genutils==0.1.0
ipywidgets==5.2.2
Jinja2==2.9.4
jsonpickle==0.9.4
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.4.0
jupyter-console==5.1.0
jupyter-core==4.3.0
Markdown==2.2.0
MarkupSafe==0.23
matlabengineforpython===R2016a
matplotlib==2.0.0
mistune==0.7.3
mock==2.0.0
munch==2.1.1
nbconvert==5.1.1
nbformat==4.2.0
notebook==4.3.1
numpy==1.12.1
nvidia-ml-py==7.352.0
olefile==0.44
packaging==16.8
pandocfilters==1.4.1
pathlib2==2.2.0
pbr==2.0.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.0.0
prompt-toolkit==1.0.9
protobuf==3.2.0
ptyprocess==0.5.1
pygit==0.1
Pygments==2.1.3
pymongo==3.4.0
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2016.10
pyzmq==16.0.2
qtconsole==4.2.1
sacred==0.7b3
scandir==1.4
scipy==0.19.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
smmap2==2.0.1
subprocess32==3.2.7
tensorflow==1.1.0rc1
terminado==0.6
testpath==0.3
tornado==4.4.2
traitlets==4.3.1
wcwidth==0.1.7
Werkzeug==0.12.1
widgetsnbextension==1.2.6
wrapt==1.10.10
